Initial Commit
This is actually pretty far into the development cycle, but its the
point at which work is worth saving.  At this point the entire flow
from the control server down to the board works, as well as the
on-board serial link.  Indicator LEDs work, and more work is happening
to make the entire system more stable and easier to use.