Add Ethernet support
Document the network connectivity flow.

We change mqtt to a pointer
	This solves the mqtt constructor needing a specific Client
	type (not the abstract one) at start time. Either we pass
	around the mqtt object everywhere, which is not presently
	how the code is architected, or we set this one toggle.

	I am not above checking for pointer equality to determine
	if we are meant to be on wifi or Ethernet for the bits that
	care about it

Ethernet DHCP happens as blocking. Sadly.

We set all the SPI pins for SPI 0 and 1, since SPI1 is only reachable on
specific pins anyway.

We do safely handle a missing wiznet device, so that's useful.

Indicators are also more correctly set.
	No need for special colors for ethernet link. A user can
	hardly miss the big blinking ethernet cable plugged in.

We make Ethernet usurp wireless
	The only other state where wireless does not pre-empt Ethernet
	is NET_CONNECT_WAIT_WIFI and that one has a timeout, so
	it's OK

mDNS, and in fact all of DNS, turned out to be innovation tokens too
far. DNS only works for the most recently #include'd interface type
(that's not great when you have both Ethernet and WiFI working on
a single boot), and mDNS was overhead that made less sense when we
start to control the entire network anyway.

We use an Administrative MAC, 02:00:00:$HI:$LO:00 for the Gizmo. This is
the high 8 bits, and the low 8 bits, of the team number as an integer.