Use fixed MQTT ID
This is the correct fix instead of the randomness introducted in
f1d44fb034.  The "Clean Session" flag tells the MQTT server that we
are a new client and don't wish to inherit any state.  This is an
explicit statement of intention, and more clearly communicates what we
expect to happen to any stale dataflows.